build(rust): stop building TF-A from RF-A repository
The developer is now expected to:
- provide the path to a valid TF-A source tree through the "TFA"
  environment variable.
- build from the rust/ directory.

So building everything (TF-A BL1 + BL2, RF-A and package everything up
in a FIP) goes like this:

  export TFA=/path/to/tfa
  cd rust
  make fvp

For the FIP, note that we fetch the fip.bin file generated by TF-A build
system, which originally contains C TF-A BL31 image. Then we replace the
BL31 image with RF-A in place inside this fip.bin file using the fiptool
(which we build from TF-A repo).

Also, delete TF-A's top-level Makefile (the one above the rust/
directory). We no longer need it, and it removes any ambiguity for the
developer about which Makefile to use.

Update the README file to reflect these changes.

Change-Id: I7089045e9b8286bcf2cde7407a172bdbd97d1032
Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>