build(rust): stop building TF-A binaries from RF-A repository
In preparation of making RF-A a standalone repository separate from
TF-A's, we now build TF-A binaries from an external TF-A source tree.
A future commit will remove all TF-A source files we do not need to
build RF-A.

The developer is now expected to:
- provide the path to a valid TF-A source tree through the "TFA"
  environment variable.
- build from the rust/ directory.

So building everything (TF-A BL1 + BL2, RF-A and package everything up
in a FIP) now goes like this:

  export TFA=/path/to/tfa
  cd rust
  make fvp

If required, build flags might be passed to TF-A build system through
the "TFA_FLAGS" environment variable. See rust/README.md for an example
of that.

For the FIP, note that we fetch the fip.bin file generated by TF-A build
system, which originally contains C TF-A BL31 image. Then we replace the
BL31 image with RF-A in place inside this fip.bin file using the fiptool
(which we build from TF-A repo).

Note that the latest version of TF-A at the time of writing does not
work with RF-A on QEMU platform. Until this is addressed, the simplest
solution is to fetch the version which RF-A was forked from, i.e. SHA1
fe4df8bdae0a5d.

Update the README file to reflect these changes.

Change-Id: I7089045e9b8286bcf2cde7407a172bdbd97d1032
Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>