feat(rust): implement warm boot entrypoint
Implement the warm boot entrypoint. This will be used in two scenarios:
when the core wakes up after CPU_SUSPEND, or after CPU_OFF (the latter
is treated the same as powering on the core for the first time).

In case of CPU_OFF, after waking up, all worlds' context should be
initialised similar to the cold boot case, but the entry points and
arguments will be different. The PSCI service provides the entry point
and arguments for NWd, and SWd should be entered at the secondary entry
point that was registered by the SPMC during init on the primary core.

After CPU_SUSPEND, execution of NWd should continue at the location and
with the arguments provided by the PSCI service. SWd execution should
continue at the previous location, with a specific FF-A message provided
by the SPMD.

Change-Id: If09c934a30c89a86f7e9d5b3af7efd3c53f83342
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>