build(rust): Make build process vendor layer compatible
The build process will work irrespective of how the platforms are
arranged. Currently it requires that all platforms are arranged in a
flat list in platorms/. This patch gives us the option to organise them
in platforms/{vendor}/{board}/. This does assume that platform names are
unique, which is fair. C TF-A also assumes the same.

Change-Id: I1ae00645711ee6e5b863368fbd1f627541941e50
Signed-off-by: Snehal Koukuntla <snehalreddy@google.com>