fix: crash reporting on FVP
Save the LR register in FVP's plat_crash_print_regs implementation, in
order to prevent the nested calls in arm_print_gic_regs to compromise
the return address of the function. plat_crash_print_regs was originally
a macro where this was not needed.
Also redirect the UART1 output to the FVP standard output.

Change-Id: I16ab0805eaf3d23092201f9aff1facb07665c986
Signed-off-by: Imre Kis <imre.kis@arm.com>