fix: avoid atomics before enabling MMU
Exclusive load/store instructions must not be used before the MMU and
caches are enabled, since they can cause unpredictable behavior.

This patch removes SpinMutex and Once from pagetable initialization, as
they rely on atomics. Instead, it introduces the VirtualMemoryMapping
struct, which provides primitives for initializing and enabling virtual
memory mapping. It mimics Once-like behavior but avoids atomics, relying
on the fact that the MMU is disabled. In this state, all memory accesses
are device memory accesses, so changes are visible to other cores.

Change-Id: I8b1351dcaa7f726429caaaf67648c7280c5a4477
Signed-off-by: Imre Kis <imre.kis@arm.com>