feat(arm-sysregs): allow different assembly sysreg name
For some system registers not recognised by the assembler, e.g.
because the assembler doesn't know about new architectural features or
because they are implementation-specific, it is necessary to use a
name like s3_0_c5_c3_1 that specifies the encoding, but we still want
to use a human-readable name for the accessor functions.

Change-Id: Id1673cf9b0566bcb57642a81a83f81254c13fda9
Signed-off-by: Andrew Walbran <qwandor@google.com>