feat(osi): add SET_SUSPEND_MODE SMC and advertise OSI
This change begins to add support for OS-initiated (OSI) mode.
OS-initiated mode is described in the PSCI spec (DEN0022F.b) and on the
following page:
https://trustedfirmware-a.readthedocs.io/en/latest/design_documents/psci_osi_mode.html.

Specifically, this change does the following: 1. It starts reporting OSI
mode support if the platform supports it. This is done through the
PSCI_FEATURES SMC as described in section 5.16.2 of the PSCI spec.

2. It implements the PSCI_SET_SUSPEND_MODE SMC call (see section 5.21 of
the PSCI spec).

Platforms can declare their support of OSI mode through the FEATURES
variable in their implementation of the PsciPlatformInterface trait.

A simple test for the SET_SUSPEND_MODE SMC has been added to STF as well.

Issue: #80
Change-Id: I20dc4c0d097af5360db0b66beafa7a64cd4d94af
Signed-off-by: Jim Ray <jimray@google.com>