feat(extensions): add remaining feats to create_spsr
Ensure that all bits of SPSR are set correctly, based on the features
present in the HW, when injecting an undef exception to lower ELs.
Reorder the code in the create_spsr() function so that the bits of SPSR
are set in the order they appear in the register to aid auditing.

Change-Id: I14f56c1c2d56c083dcfc2fb5c040dbb5a11f620c
Signed-off-by: Jamie Fox <jamiefox@google.com>