feat(rme): add payload for R-EL2 & build instructions
Introduces a new STF image which runs in the Realm World, instantly
returning to the Root World with the SMC indicating a successful R-EL2
boot.

Realm worlds requires three extensions to be present: FGT2, TCR2 and
HCX. This patch adds them to the FVP cpu extensions and along with the
implementation for FGT2.

Finally, this also updates the FVP arguments in `build-and-run.sh` when
RME is enabled, based on the shrinkwrap configuration for running 4
worlds.

Change-Id: Ifc97c51d722d991bfaa1da1708aae8cceb9b1c8c
Signed-off-by: Ludovic Mermod <ludovic.mermod@arm.com>