feat(extensions): enable FEAT_PAuth at lower ELs
Enables PAuth at lower ELs and adds support for saving and restoring the
PAuth key registers. The registers are saved & restored in assembly
before executing Rust code on EL3 entry and afterwards on EL3 exit so
that PAuth can be enabled at EL3 in the future.

As FEAT_PAuth is mandatory from Armv8.3 it is assumed to be present and
widely used by lower ELs, so the PAuth context is added unconditionally.

Change-Id: Ic8d5a8ac82bdfb8aa7e480c01ba57c4d6c94ffc7
Signed-off-by: Jamie Fox <jamiefox@google.com>