test: improve output of STF
This makes the output of normal world and secure world tests more
consistent, makes the messages for each test line up, and moves the
summary of normal world test counts to the end so that it can more
easily be found.

Also reduced the level of some verbose log messages.

Change-Id: I4593c204676e4f65259dab91c7a63ef44081b345
Signed-off-by: Andrew Walbran <qwandor@google.com>